Crate procinfo

procinfo is a simple program to demonstrate the use of the libproc library functions.

It prints out info about a process specified by its pid, or the current process if no pid specified.

Usage

 procinfo [pid]

Which will produce output similar to:

Libversion: 1.1
Pid: 8484
Path: /Users/andrew/workspace/libproc-rs/target/debug/procinfo
Name: procinfo
Region Filename (at address 0): /Users/andrew/workspace/libproc-rs/target/debug/procinfo
There are currently 454 processes active
8496
...

Modules

c 🔒

Functions

getpid 🔒

main 🔒

procinfo 🔒