Struct procinfo::pid::Status

pub struct Status {
    pub command: String,
    pub state: State,
    pub pid: pid_t,
    pub numa_gid: pid_t,
    pub tid: pid_t,
    pub ppid: pid_t,
    pub tracer_pid: pid_t,
    pub uid_real: uid_t,
    pub uid_effective: uid_t,
    pub uid_saved: uid_t,
    pub uid_fs: uid_t,
    pub gid_real: gid_t,
    pub gid_effective: gid_t,
    pub gid_saved: gid_t,
    pub gid_fs: gid_t,
    pub fd_allocated: u32,
    pub groups: Vec<gid_t>,
    pub ns_pids: Vec<pid_t>,
    pub ns_tids: Vec<pid_t>,
    pub ns_pgids: Vec<pid_t>,
    pub ns_sids: Vec<pid_t>,
    pub vm_peak: usize,
    pub vm_size: usize,
    pub vm_locked: usize,
    pub vm_pin: usize,
    pub vm_hwm: usize,
    pub vm_rss: usize,
    pub vm_data: usize,
    pub vm_stack: usize,
    pub vm_exe: usize,
    pub vm_lib: usize,
    pub vm_pte: usize,
    pub vm_pmd: usize,
    pub vm_swap: usize,
    pub threads: u32,
    pub sig_queued: u64,
    pub sig_queued_max: u64,
    pub sig_pending_thread: u64,
    pub sig_pending_process: u64,
    pub sig_blocked: u64,
    pub sig_ignored: u64,
    pub sig_caught: u64,
    pub cap_inherited: u64,
    pub cap_permitted: u64,
    pub cap_effective: u64,
    pub cap_bounding: u64,
    pub seccomp: SeccompMode,
    pub cpus_allowed: Box<[u8]>,
    pub mems_allowed: Box<[u8]>,
    pub voluntary_ctxt_switches: u64,
    pub nonvoluntary_ctxt_switches: u64,
}

Process status information.

See man 5 proc and Linux/fs/proc/array.c.

Fields

command

Filename of the executable.

state

Current state of the process.

pid

Process ID (i.e., Thread Group ID).

numa_gid

NUMA group ID.

tid

Thread ID.

ppid

Process ID of parent process.

tracer_pid

Process ID of the process tracing this process (0 if not being traced).

uid_real

Real user ID.

uid_effective

Effective user ID.

uid_saved

Saved user ID.

uid_fs

Filesystem user ID.

gid_real

Real group ID.

gid_effective

Effective group ID.

gid_saved

Saved group ID.

gid_fs

Filesystem group ID.

fd_allocated

Number of file descriptor slots currently allocated.

groups

Supplementary group list.

ns_pids

Process IDs for each namespace which the process belongs to.

ns_tids

Thread IDs for each namespace which the process belongs to.

ns_pgids

Process group IDs for each namespace which the process belongs to.

ns_sids

Session IDs of the process for each namespace to which it belongs.

vm_peak

Peak virtual memory size (kB).

vm_size

Virtual memory size (kB).

vm_locked

Locked memory size (kB) (see mlock(3)).

vm_pin

Pinned memory size (since Linux 3.2). These are pages that can't be moved because something needs to directly access physical memory.

vm_hwm

Peak resident size (kB) ("high water mark").

vm_rss

Resident set size.

vm_data

Size of data segments (kB).

vm_stack

Size of stack segments (kB).

vm_exe

Size of text (executable) segments (kB).

vm_lib

Shared library code size (kB).

vm_pte

Page table entries size (since Linux 2.6.10).

vm_pmd

Size of second-level page tables (since Linux 4.0).

vm_swap

Swapped-out-virtual memory size (since Linux 2.6.34).

threads

Number of threads in process containing this thread.

sig_queued

The number of currently queued signals for this real user ID (see the description of RLIMIT_SIGPENDING in getrlimit(2)).

sig_queued_max

The resource limit on the number of queued signals for this process.

sig_pending_thread

Number of signals pending for the thread (see pthreads(7)).

sig_pending_process

Number of signals pending for the process (see signal(7)).

sig_blocked

Mask indicating signals being blocked.

sig_ignored

Mask indicating signals being ignored.

sig_caught

Mask indicating signals being caught.

cap_inherited

Mask of capabilities enabled in inheritable sets (see capabilities(7)).

cap_permitted

Mask of capabilities enabled in permitted sets.

cap_effective

Mask of capabilities enabled in effective sets.

cap_bounding

Capability Bounding set (since Linux 2.6.26).

seccomp

Secure Computing mode of the process (since Linux 3.8, see seccomp(2)). This field is provided only if the kernel was built with the CONFIG_SECCOMP kernel configuration option enabled.

cpus_allowed

CPUs on which this process may run (since Linux 2.6.24, see cpuset(7)).

The slice represents a bitmask in the same format as BitVec.

mems_allowed

Memory nodes allowed to this process (since Linux 2.6.24, see cpuset(7)).

The slice represents a bitmask in the same format as BitVec.

voluntary_ctxt_switches

Number of voluntary context switches.

nonvoluntary_ctxt_switches

Number of involuntary context switches.