Struct procinfo::LoadAvg

pub struct LoadAvg {
    pub load_avg_1_min: f32,
    pub load_avg_5_min: f32,
    pub load_avg_10_min: f32,
    pub tasks_runnable: u32,
    pub tasks_total: u32,
    pub last_created_pid: pid_t,
}

System load and task statistics.

The load average is the ratio of runnable and uninterruptible (waiting on IO) tasks to total tasks on the system.

See man 5 proc and Linux/fs/proc/loadavg.c.

Fields

load_avg_1_min

Load average over the last minute.

load_avg_5_min

Load average of the last 5 minutes.

load_avg_10_min

Load average of the last 10 minutes

tasks_runnable

the number of currently runnable kernel scheduling entities (processes, threads).

tasks_total

the number of kernel scheduling entities that currently exist on the system.

last_created_pid

the PID of the process that was most recently created on the system.