Module procinfo::pid

Process-specific information from /proc/[pid]/.

Structs

Cpu

Io
The I/O information of a process

Limit
A resource limit, including a soft and hard bound.

Limits
Process limits information See man 2 getrlimit.

Mountinfo
Process mounts information.

Stat
Process status information.

Statm
Process memory usage information.

Status
Process status information.

Enums

SeccompMode
The Secure Computing state of a process.

State
The state of a process.

Functions

cwd
Gets path of current working directory for the process with the provided pid.

cwd_self
Gets path of current working directory for the current process.

io
Returns I/O information for the process with the provided pid.

io_self
Returns I/O information for the current process.

io_task
Returns I/O information from the thread with the provided parent process ID and thread ID.

limits
Returns resource limit information from the process with the provided pid.

limits_self
Returns resource limit information for the current process.

limits_task
Returns resource limit information from the thread with the provided parent process ID and thread ID.

mountinfo
Returns mounts information for the process with the provided pid.

mountinfo_self
Returns mounts information for the current process.

mountinfo_task
Returns mounts information from the thread with the provided parent process ID and thread ID.

stat
Returns status information for the process with the provided pid.

stat_self
Returns status information for the current process.

stat_task
Returns status information from the thread with the provided parent process ID and thread ID.

statm
Returns memory status information for the process with the provided pid.

statm_self
Returns memory status information for the current process.

statm_task
Returns memory status information from the thread with the provided parent process ID and thread ID.

status
Returns memory status information for the process with the provided pid.

status_self
Returns memory status information for the current process.

status_task
Returns memory status information from the thread with the provided parent process ID and thread ID.